move read_product_metadata to utils
